backend/requirements.txt
# FastAPI + SQLAlchemy + Pydantic
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.0

# JWT 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 이미지 처리
albumentations==1.3.1
opencv-python-headless==4.9.0.80
pillow>=11.0.0
numpy>=1.24.0,<2.0.0
segmentation-models-pytorch==0.3.3

# 유틸리티
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0

# PyTorch - 별도 설치 필요
# GPU (CUDA 12.1): uv pip install torch==2.5.1 torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cu121
# CPU: uv pip install torch==2.5.1 torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cpu
